refactor: extract virtual XSL template handling into common.xsl package
